Changes in archiving for DATETIME id in backups

+++ usr/share/drls/backup/run/default/059_prepare_backup_preservation.sh
@@ -5,25 +5,33 @@ if [ -d ${PXEDIR}/${CLINAME} ]
 then
         if [ -d ${PXEDIR}/${CLINAME}/.archive ]
         then
-                PXEDATE=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $1}')
-                tar --exclude='.archive' -C ${PXEDIR}/${CLINAME} . -cf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.$PXEDATE.pxe.arch
-		if [ $? -eq 0 ]
+		if [ -f ${PXEDIR}/${CLINAME}/${CLINAME}.kernel ]
 		then
-                	rm -vf ${PXEDIR}/${CLINAME}/.lockfile ${PXEDIR}/${CLINAME}/*
-		else
-			rm -vf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.$PXEDATE.pxe.arch
-			StopIfError "Problem archiving previous DR imagebackup. See log ${LOGFILE} for details"
+                 	PXETIME=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $2}' | awk -F"." '{print $1}' | tr -d ":")
+                	PXEDATE=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $1}' | tr -d "-")
+                	tar --exclude='.archive' -C ${PXEDIR}/${CLINAME} . -cf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.${PXEDATE}${PXETIME}.pxe.arch
+			if [ $? -eq 0 ]
+			then
+                		rm -vf ${PXEDIR}/${CLINAME}/.lockfile ${PXEDIR}/${CLINAME}/*
+			else
+				rm -vf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.$PXEDATE.pxe.arch
+				StopIfError "Problem archiving previous DR imagebackup. See log ${LOGFILE} for details"
+			fi
 		fi
         else
-                mkdir -v ${PXEDIR}/${CLINAME}/.archive
-                PXEDATE=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $1}')
-                tar --exclude='.archive' -C ${PXEDIR}/${CLINAME} . -cf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.$PXEDATE.pxe.arch
-		if [ $? -eq 0 ]
+		if [ -f ${PXEDIR}/${CLINAME}/${CLINAME}.kernel ]
 		then
-                	rm -vf ${PXEDIR}/${CLINAME}/.lockfile ${PXEDIR}/${CLINAME}/*
-		else
-			rm -vf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.$PXEDATE.pxe.arch
-			StopIfError "Problem archiving previous DR imagebackup. See log ${LOGFILE} for details"
+                	mkdir -v ${PXEDIR}/${CLINAME}/.archive
+                 	PXETIME=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $2}' | awk -F"." '{print $1}' | tr -d ":")
+                	PXEDATE=$(stat -c %y ${PXEDIR}/${CLINAME}/${CLINAME}.kernel | awk '{print $1}' | tr -d "-")
+                	tar --exclude='.archive' -C ${PXEDIR}/${CLINAME} . -cf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.${PXEDATE}${PXETIME}.pxe.arch
+			if [ $? -eq 0 ]
+			then
+                		rm -vf ${PXEDIR}/${CLINAME}/.lockfile ${PXEDIR}/${CLINAME}/*
+			else
+				rm -vf ${PXEDIR}/${CLINAME}/.archive/${CLINAME}.${PXEDATE}${PXETIME}.pxe.arch
+				StopIfError "Problem archiving previous DR imagebackup. See log ${LOGFILE} for details"
+			fi
 		fi
         fi
 #else
@@ -34,27 +42,35 @@ if [ -d ${BKPDIR}/${CLINAME} ]
 then
         if [ -d ${BKPDIR}/${CLINAME}/.archive ]
         then
-                BKPDATE=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $1}')
-                tar --exclude='.archive' -C ${BKPDIR}/${CLINAME} . -cf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.$BKPDATE.bkp.arch
-		if [ $? -eq 0 ]
+		if [ -f ${BKPDIR}/${CLINAME}/backup.tar.gz ]
 		then
-                	rm -vf ${BKPDIR}/${CLINAME}/*
-		else
-			rm -vf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.$BKPDATE.bkp.arch
-			StopIfError "Problem archiving previous DR backup. See log ${LOGFILE} for details"
+                	BKPTIME=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $2}' | awk -F"." '{print $1}' | tr -d ":")
+                	BKPDATE=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $1}' | tr -d "-")
+                	tar --exclude='.archive' -C ${BKPDIR}/${CLINAME} . -cf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.${BKPDATE}${BKPTIME}.bkp.arch
+			if [ $? -eq 0 ]
+			then
+                		rm -vf ${BKPDIR}/${CLINAME}/*
+			else
+				rm -vf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.${BKPDATE}${BKPTIME}.bkp.arch
+				StopIfError "Problem archiving previous DR backup. See log ${LOGFILE} for details"
+			fi
 		fi
         else
-                mkdir -v ${BKPDIR}/${CLINAME}/.archive
-                BKPDATE=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $1}')
-                tar --exclude='.archive' -C ${BKPDIR}/${CLINAME} . -cf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.$BKPDATE.bkp.arch
-		if [ $? -eq 0 ]
+		if [ -f ${BKPDIR}/${CLINAME}/backup.tar.gz ]
 		then
-                	rm -vf ${BKPDIR}/${CLINAME}/*
-		else
-			rm -vf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.$BKPDATE.bkp.arch
-			StopIfError "Problem archiving previous DR backup. See log ${LOGFILE} for details"
-		fi
-        fi
+                	mkdir -v ${BKPDIR}/${CLINAME}/.archive
+                	BKPTIME=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $2}' | awk -F"." '{print $1}' | tr -d ":")
+                	BKPDATE=$(stat -c %y ${BKPDIR}/${CLINAME}/backup.tar.gz | awk '{print $1}' | tr -d "-")
+                	tar --exclude='.archive' -C ${BKPDIR}/${CLINAME} . -cf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.${BKPDATE}${BKPTIME}.bkp.arch
+			if [ $? -eq 0 ]
+			then
+                		rm -vf ${BKPDIR}/${CLINAME}/*
+			else
+				rm -vf ${BKPDIR}/${CLINAME}/.archive/${CLINAME}.${BKPDATE}${BKPTIME}.bkp.arch
+				StopIfError "Problem archiving previous DR backup. See log ${LOGFILE} for details"
+			fi
+        	fi
+	fi
 #else
         #do nothing...
 fi
